requirements.txt
asgiref==3.11.0
certifi==2026.1.4
charset-normalizer==3.4.4
Django==6.0.1
django-cors-headers==4.9.0
django-environ==0.12.0
django-filter==25.2
django-redis==6.0.0
django-rest-passwordreset==1.5.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.11
pillow==12.1.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-dotenv==1.2.1
razorpay==2.0.0
redis==7.1.0
requests==2.32.5
sqlparse==0.5.5
stripe==14.3.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
celery
gunicorn
argon2-cffi
pytest
pytest-django
factory-boy
pytest-cov
bandit